requirements.txt
requests>=2.31
jinja2>=3.1
google-genai>=1.0
